requirements.txt
discord.py>=2.3.2
filelock>=3.12.4
